backend/dependencies/UsersHub-authentification-module/requirements.txt
bcrypt
flask-sqlalchemy>=2.3.2
flask>=1.1.1
psycopg2>=2.8.5
requests>=2.25.0
sqlalchemy>=1.3.19
utils-flask-sqlalchemy
flask-marshmallow>=0.12.0
